Use MbfTable for Registered devices and Clusters tables
