+++ frontend/package.json
@@ -32,16 +32,10 @@
     "install:qrcode": "npm install qrcode.react --omit=dev",
     "install:notistack": "npm install notistack --omit=dev",
     "install:rjsf": "npm install @rjsf/core @rjsf/utils @rjsf/validator-ajv8 --omit=dev",
-    "install:react-table": "npm install react-table@latest --omit=dev && npm install --save-dev @types/react-table",
     "install:typescript": "npm install --save-dev typescript @types/node",
     "install:vite": "npm install --save-dev vite @vitejs/plugin-react",
     "install:vitest": "npm install --save-dev vitest jsdom @testing-library/react @testing-library/jest-dom",
     "install:eslint": "npm install --save-dev eslint eslint-plugin-react eslint-plugin-react-hooks typescript-eslint @typescript-eslint/eslint-plugin @typescript-eslint/parser"
-  },
-  "overrides": {
-    "react-table": {
-      "react": "^19.0.0"
-    }
   },
   "dependencies": {
     "@emotion/react": "^11.14.0",
@@ -58,8 +52,7 @@
     "react": "^19.1.1",
     "react-dom": "^19.1.1",
     "react-router": "^7.9.1",
-    "react-router-dom": "^7.9.1",
-    "react-table": "^7.8.0"
+    "react-router-dom": "^7.9.1"
   },
   "devDependencies": {
     "@eslint/js": "^9.35.0",
